+++ src/packages/components/src/functions/index.ts
@@ -1,2 +1,3 @@
+export { debounce } from './debounce'
 export { colSpanClass } from './tailwind'
 export type { GridColSpan } from './tailwind'